pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "carmen"
version = "0.1.0"
description = "Measure and report carbon footprint of Kubernetes applications and cloud infrastructure"
readme = "README.md"
license = {text = "MIT"}
keywords = [
    "carbon",
    "emissions",
    "sustainability",
    "kubernetes",
    "prometheus",
    "thanos",
    "azure",
    "green-software",
    "energy-efficiency",
    "cloud-computing",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Monitoring",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Framework :: Pydantic",
    "Framework :: Pydantic :: 2",
]
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.104.1,<0.105.0",
    "uvicorn[standard]>=0.24.0,<0.25.0",
    "pydantic>=2.5.0,<2.6.0",
    "pydantic-settings>=2.1.0,<2.2.0",
    "starlette>=0.27.0,<0.28.0",
    "PyYAML>=6.0.1,<7.0.0",
    "requests>=2.31.0,<3.0.0",
    "urllib3>=2.1.0,<3.0.0",
    "httpx>=0.26.0,<0.27.0",
    "msal>=1.26.0,<2.0.0",
    "azure-storage-blob>=12.19.0,<13.0.0",
    "azure-identity>=1.15.0,<2.0.0",
    "azure-core>=1.31.0,<2.0.0",
    "pandas>=2.2.2,<3.0.0",
    "pyarrow>=16.1.0,<17.0.0",
    "numpy>=2.2.0,<3.0.0",
    "colorlog>=6.8.2,<7.0.0",
    "Jinja2>=3.1.2,<4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-xdist>=2.5.0",
    "coverage>=7.4.4",
    "pylint>=3.2.6",
    "httpretty>=1.1.4",
]

test = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-xdist>=2.5.0",
    "coverage>=7.4.4",
    "httpretty>=1.1.4",
    "robotframework>=7.0",
    "robotframework-requests>=0.9.7",
]

all = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-xdist>=2.5.0",
    "coverage>=7.4.4",
    "pylint>=3.2.6",
    "httpretty>=1.1.4",
    "robotframework>=7.0",
    "robotframework-requests>=0.9.7",
]

[project.urls]

[project.scripts]
carbon-engine = "backend.src.main:main"
carbon-daemon = "backend.src.daemon.carbon_daemon:main"

[tool.setuptools]
zip-safe = false

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
exclude = [
    "*.tests",
    "*.tests.*",
    "tests.*",
    "tests",
    "*.__pycache__",
    "*.egg-info",
    ".venv"
]

[tool.setuptools.package-data]
backend = [
    "py.typed",
    "**/*.yaml",
    "**/*.yml",
    "**/*.yml.j2",
    "**/*.json",
    "**/*.csv",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--asyncio-mode=auto",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"
